BA2 all solved

+++ makepy.py
@@ -1,8 +1,11 @@
-# 파일 생성 반복 코드
-prefix = "BA2"
-n = 8
+import os
+
+prefix = "BA3"
+n = 0
 ASCII_big = [chr(i) for i in range(65, 65+n)]
+
 for j in range(n):
     filename = prefix + ASCII_big[j] + ".py"
-    with open(filename,"w") as f:
-        f.write("")
+    if not os.path.exists(filename):
+        with open(filename, "w") as f:
+            f.write("")
